backend/requirements.txt
Flask==2.2.3
connexion[swagger-ui]
Flask-Injector
pycryptodome
pyodbc
ConfigParser
xmltodict
python-dateutil
defusedxml
mysql-connector-python
python-dotenv
pytest
mock
autopep8
coverage
pytest-cov
pytest-html
mysql-connector-python